requirements_dev.txt

#Coverage
coverage==7.4.0
coverage-badge==1.1.0

# DVC
dvc >= 3.38.1
dvc[gs]

fastapi
hydra-core

# docs
mkdocs <= 1.5.3

# docs
mkdocs <= 1.5.3
mkdocs-material <= 9.4.6
mkdocs-material <= 9.4.6
mkdocstrings <= 0.23.0
mkdocstrings <= 0.23.0
mkdocstrings[python]>=0.18
mkdocstrings[python]>=0.18
monai
pillow
# pre-commit
pre-commit >= 3.5.0

# testing
pytest >= 7.4.3
python-multipart
pytorch-lightning==2.1.3

# linting
ruff >= 0.1.3
tensorboard

torch >= 2.1.0
torchtest==0.5
torchvision
uvicorn[standard]

python-multipart
torchvision